# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TechFuturist API Documentation",
        "description": "![TechFuturist Logo](http://localhost:8000/images/images.png)\n * Documentation de l'API pour l'application TechFuturist",
        "contact": {
            "email": "support@techfuturist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Récupère toutes les ressources avec statistiques (Admin uniquement)",
                "operationId": "6ec1827c29e12a5567ba377df35a82ec",
                "responses": {
                    "200": {
                        "description": "Liste des ressources avec statistiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resources": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Resource"
                                            }
                                        },
                                        "total_resources": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "published_resources": {
                                            "type": "integer",
                                            "example": 80
                                        },
                                        "total_downloads": {
                                            "type": "integer",
                                            "example": 2500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Crée une nouvelle ressource (Admin uniquement)",
                "operationId": "846d6115fa12699a52e27b87bc05cb83",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "file_type",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Guide Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un guide complet sur Laravel 10"
                                    },
                                    "file_type": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "pdf"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ressource créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resources/{id}": {
            "put": {
                "tags": [
                    "Resources"
                ],
                "summary": "Met à jour une ressource (Admin uniquement)",
                "operationId": "1f0b00850bbb49af2def76e08835b076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Guide Laravel mis à jour"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mise à jour du guide avec les dernières fonctionnalités"
                                    },
                                    "file_type": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "pdf"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ressource mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Resources"
                ],
                "summary": "Supprime une ressource (Admin uniquement)",
                "operationId": "41263949cef614080b6337bb44e32176",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ressource supprimée avec succès"
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resources/download/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Télécharge un fichier ressource",
                "operationId": "c39eee02efcab3dd17456e73e19dac5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier téléchargé"
                    },
                    "404": {
                        "description": "Ressource ou fichier non trouvé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Resource": {
                "title": "Resource",
                "required": [
                    "id",
                    "title",
                    "file_url",
                    "file_type",
                    "uploaded_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Guide Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un guide complet sur Laravel 10"
                    },
                    "file_url": {
                        "type": "string",
                        "example": "uploads/abc123.pdf"
                    },
                    "file_type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "uploaded_by": {
                        "type": "integer",
                        "example": 2
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": true
                    },
                    "download_count": {
                        "type": "integer",
                        "example": 15
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-08T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-10T15:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}